Netdata

+++ .sync/Archive/routes/web.1.php
@@ -17,32 +17,32 @@ Route::get('about', 'PagesController@getAbout');
 
 // Route to contact page
 Route::get('contact', 'PagesController@getContact');
-Route::post('contact', 'PagesController@postContact'); // <<<< MUST FINISH
+Route::post('contact', 'PageController@postContact'); // <<<< MUST FINISH
 
 
 // Route to generated dynamic pages
-Route::get('news', 'NewsController@getNews');
-Route::get('news/{slug}', ['as' => 'news/post', 'uses' => 'NewsController@getPost'])->where('slug', '[\w\d\-\_]+');
+Route::get('news', 'BlogController@getBlog');
+Route::get('news/{slug}', ['as' => 'news/post', 'uses' => 'BlogController@getPost'])->where('slug', '[\w\d\-\_]+');
 
 
 // Route to admin area
 Route::get('admin', 'AdminController@getAdmin');
 
 // Route to dynamic Post pages
-Route::resource('posts', 'PostController');
+Route::resource('admin/posts', 'PostController');
 
 // Route to Categories
-Route::resource('categories', 'CategoryController', ['except' => ['create']]);
+Route::resource('admin/categories', 'CategoryController', ['except' => ['create']]);
 
 // Route to Tags
-Route::resource('tags', 'TagController', ['except' => ['create']]);
+Route::resource('admin/tags', 'TagController', ['except' => ['create']]);
 
 // Route to Comments
-Route::post('comments/{post_id}', ['uses' => 'CommentsController@store', 'as' => 'comments.store'] );
-Route::get('comments/{id}/edit', ['uses' => 'CommentsController@edit', 'as' => 'comments.edit'] );
-Route::put('comments/{id}', ['uses' => 'CommentsController@update', 'as' => 'comments.update'] );
-Route::delete('comments/{id}', ['uses' => 'CommentsController@destroy', 'as' => 'comments.destroy'] );
-Route::get('comments/{id}/delete', ['uses' => 'CommentsController@delete', 'as' => 'comments.delete'] );
+Route::post('admin/comments/{post_id}', ['uses' => 'CommentsController@store', 'as' => 'admin/comments/store'] );
+Route::get('admin/comments/{id}/edit', ['uses' => 'CommentsController@edit', 'as' => 'admin/comments/edit'] );
+Route::put('admin/comments/{id}', ['uses' => 'CommentsController@update', 'as' => 'admin/comments/update'] );
+Route::delete('admin/comments/{id}', ['uses' => 'CommentsController@destroy', 'as' => 'admin/comments/destroy'] );
+Route::get('admin/comments/{id}/delete', ['uses' => 'CommentsController@delete', 'as' => 'admin/comments/delete'] );
 
 
 Auth::routes();
